+++ test/google/cloud/firestore/query_test.rb
@@ -11,18 +11,6 @@ module Google
           @book_3 = Book.create!(title: 'My 3rd book', published_on: '2022-12-03'.to_date, page: 230)
         end
 
-        test '#get_records' do
-          books = Book.where(:page, :>=, 220).order(:page)
-          book_ids = [@book_2, @book_3].map(&:id)
-          found_ids = books.map(&:id)
-          assert_equal book_ids, found_ids
-
-          books = Book.where(:page, :>=, 220).limit(1)
-          assert_equal 1, books.size
-          found_id = books[0].id
-          assert_includes book_ids, found_id
-        end
-
         test '#destroy_all' do
           Book.where(:page, :>=, 220).destroy_all