+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Next
+## v0.0.10 - 2022-12-28
 - Implement `inspect` and `pretty_print` for Google::Cloud::Firestore::Query.
 ## v0.0.9 - 2022-12-27
 - Remove `get_records` method. Now records can be retrieved without it.